chore(client): Move stardust landing files to own folder + Add minor code structure refactor
