+++ client/src/app/routes/stardust/landing/Landing.scss
@@ -1,8 +1,8 @@
-@import "../../../scss/fonts";
-@import "../../../scss/mixins";
-@import "../../../scss/media-queries";
-@import "../../../scss/variables";
-@import "../../../scss/themes";
+@import "../../../../scss/fonts";
+@import "../../../../scss/mixins";
+@import "../../../../scss/media-queries";
+@import "../../../../scss/variables";
+@import "../../../../scss/themes";
 
 .landing-stardust {
     display: flex;
